removed ner_duckling
