+++ tests/base/test_extractors.py
@@ -123,7 +123,7 @@ def test_crf_json_from_non_BILOU(spacy_nlp, ner_crf_pos_feature_config):
 def test_duckling_entity_extractor(component_builder):
     _config = RasaNLUModelConfig({"pipeline": [{"name": "ner_duckling"}]})
     _config.set_component_attr("ner_duckling", dimensions=["time"])
-    duckling = component_builder.create_component("ner_duckling", _config)
+    duckling = component_builder.create_component("ner_duckling_http", _config)
     message = Message("Today is the 5th of May. Let us meet tomorrow.")
     duckling.process(message)
     entities = message.get("entities")
@@ -143,7 +143,7 @@ def test_duckling_entity_extractor(component_builder):
 def test_duckling_entity_extractor_and_synonyms(component_builder):
     _config = RasaNLUModelConfig({"pipeline": [{"name": "ner_duckling"}]})
     _config.set_component_attr("ner_duckling", dimensions=["number"])
-    duckling = component_builder.create_component("ner_duckling", _config)
+    duckling = component_builder.create_component("ner_duckling_http", _config)
     synonyms = component_builder.create_component("ner_synonyms", _config)
     message = Message("He was 6 feet away")
     duckling.process(message)
